btrfs-progs: libbtrfsutil: use package_data member for header files
[BUG]
Currently with python3.12, the python bindding will always result the
following warning:

    [PY]     libbtrfsutil
/usr/lib/python3.12/site-packages/setuptools/_distutils/extension.py:134: UserWarning: Unknown Extension options: 'headers'
  warnings.warn(msg)

[CAUSE]
In the setup.py which specifies the files to be included into the
package, we use setuptools::Extension to specify the file lists and
include paths.

But there is no handling of Extension::headers member, thus resulting
the above warning.

[FIX]
Not an expert in python, but other packages like cffi are all using
`package_data` to include the headers, so just follow them to use
`package_data` member for headers.

Fixes: 87c3fb557ef2 ("libbtrfsutil: update python setup.py for distribution")
Signed-off-by: Qu Wenruo <wqu@suse.com>