btrfs-progs: remove block group free space
In the upstream equivalent of btrfs_remove_block_group(), the function
remove_block_group_free_space() is called to delete free spaces
associated with the block group being freed. However, this function is
defined in btrfs-progs but not called anywhere.

To address this issue, I added a call to remove_block_group_free_space()
in btrfs_remove_block_group(). This ensures that the free spaces are
properly deleted when a block group is removed.

I also added a check to remove_block_group_free_space to make sure that
free-space-tree is enabled.

Signed-off-by: Leo Martins <loemra.dev@gmail.com>