btrfs-progs: mkfs: add "--blocksize" option as an alias for "--sectorsize"
Btrfs uses the terminology "sectorsize" for the minimum data allocation
size, which is very different from other filesystems, and can be a
little confusing with the kernel block io sector size (which is the
minimum block io size, and is fixed to 512 bytes).

Let's start migrating mkfs.btrfs and its documentation to use the more
common terminology instead.

Signed-off-by: Qu Wenruo <wqu@suse.com>