btrfs-progs: build: add target to generate compile_commands.json
To support LSP in editors add a convenience command to generate the
commands that eg. clangd uses to generate it's database. This requires
the Bear tool to scan build. Compilation can be done by gcc or clang,
only the .json needs the explicit CC=clang. LSP is sensitive to the
CFLAGS.

Signed-off-by: David Sterba <dsterba@suse.com>