btrfs-progs: mkfs: set BTRFS_INODE_ROOT_ITEM_INIT on new FS root
The kernel commit 08fe4db170b41 from 2011 sets the flag
BTRFS_INODE_ROOT_ITEM_INIT on root items, to work around a bug where
flags and byte_limit weren't being set.

Copy this behaviour in mkfs, to prevent the kernel from having to do it
on the first mount. We memset the btrfs_root_item, so there's no
corruption issue as there once was. We already do this in
btrfs_make_subvolume(), as otherwise the readonly flag of any subvolumes
would get reset.

Signed-off-by: Mark Harmstone <maharmstone@fb.com>