btrfs-progs: unify init sequence of standalone utilities
Each utility that starts from main() (and is considered commonly used)
should init fastest checksum implementation and set up config. This was
properly done only in btrfs.c, fix it in btrfs-image, btrfs-convert and
btrfstune as all of them do a lot of checksum calculations. Other
standalone utilities do not need to do that.

- cpu_detect_flags();
- hash_init_accel();
- btrfs_config_init();

Signed-off-by: David Sterba <dsterba@suse.com>