btrfs-progs: qgroup show: use sysfs to detect inconsistent state early
Currently if "btrfs qgroup show" detects the qgroup is in an
inconsistent state, it will print a warning:

  WARNING: qgroup data inconsistent, rescan recommended

But the detection is based on the tree search ioctl, and qgroup tree is
only updated at transaction commit time.

This means if qgroup is marked inconsistent, and the transaction is not
commit, there can be a time window as long as 30s before "btrfs qgroup
show" gives a proper warning about inconsistent qgroup numbers.

To address this, use the
'/sys/fs/btrfs/<fsid>/qgroup/inconsistent' file to do extra check.

That file is updated at real time, thus there is no delay, and can give
an early warning about inconsistent qgroup numbers.

Link: https://bugzilla.suse.com/show_bug.cgi?id=1235765
Signed-off-by: Qu Wenruo <wqu@suse.com>
Signed-off-by: David Sterba <dsterba@suse.com>