btrfs-progs: check: fix false alert on missing csum for hole
If we log a hole, fsync a file, partially write to the front of the hole
and then fsync again the file, we end up with a file extent item in the
log tree that represents a hole and has a disk address of 0 and an offset
that is greater than zero (since we trimmed the front part of the range to
accomodate for a file extent item of the new write).

When this happens 'btrfs check' incorrectly reports that a csum is missing
like this:

  $ btrfs check /dev/sdc
  Opening filesystem to check...
  Checking filesystem on /dev/sdc
  UUID: 46a85f62-4b6e-4aab-bfdc-f08d1bae9e08
  [1/8] checking log
  ERROR: csum missing in log (root 5 inode 262 offset 5959680 address 0x5a000 length 1347584)
  ERROR: errors found in log
  [2/8] checking root items
  (...)

And in the log tree, the corresponding file extent item:

  $ btrfs inspect-internal dump-tree /dev/sdc
  (...)
        item 38 key (262 EXTENT_DATA 5959680) itemoff 1796 itemsize 53
                generation 11 type 1 (regular)
                extent data disk byte 0 nr 0
                extent data offset 368640 nr 1347584 ram 1716224
                extent compression 0 (none)
  (...)

This false alert happens because we sum the file extent item's offset to
its logical address before we attempt to skip holes at
check_range_csummed(), so we end up passing a non-zero logical address to
that function (0 + 368640), which will attempt to find a csum for that
invalid address and fail.

This type of error can be sporadically triggered by several test cases
from fstests such as btrfs/192 for example.

Fix this by skipping csum search if the file extent item's logical disk
address is 0 before summing the offset.

Fixes: 88dc309aca10 ("btrfs-progs: check: explicit holes in log tree don't get csummed")
Reviewed-by: Boris Burkov <boris@bur.io>
Reviewed-by: Qu Wenruo <wqu@suse.com>
Signed-off-by: Filipe Manana <fdmanana@suse.com>
Signed-off-by: David Sterba <dsterba@suse.com>