btrfs-progs: rescue: add new command fix-data-checksum
It's a long known problem that direct IO can lead to data checksum
mismatches if the user space is also modifying its buffer during
writeback.

Although it's being fixed in the recent v6.15 release (and backported to
older kernels), we still need a user friendly way to fix those problems.

This patch introduces the dryrun version of "btrfs rescue
fix-data-checksum", which reports the logical bytenr and corrupted
mirrors.

Signed-off-by: Qu Wenruo <wqu@suse.com>
Signed-off-by: David Sterba <dsterba@suse.com>