btrfs-progs: fix-data-checksum: show affected files
Previously "btrfs rescue fix-data-checksum" only showed affected logical
bytenr, which is not helpful to determine which files are affected.

Enhance the output by also printing the affected subvolumes (in numeric
form), and the file paths inside that subvolume.

The example looks like this:

  logical=13631488 corrtuped mirrors=1 affected files:
    (subvolume 5)/foo
    (subvolume 5)/dir/bar
  logical=13635584 corrtuped mirrors=1 affected files:
    (subvolume 5)/foo
    (subvolume 5)/dir/bar

Although the end result is still not perfect, it's much easier to find
out which files are affected.

Signed-off-by: Qu Wenruo <wqu@suse.com>
Signed-off-by: David Sterba <dsterba@suse.com>